dspy-samples/pyproject.toml
[tool.poetry]
name = "dspy-samples"
version = "0.1.0"
description = ""
authors = ["Gavin Noronha <gavinln@hotmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
dspy-ai = "2.4.10"
jupyterlab = "^4.2.3"
openai = "^1.35.10"
ipywidgets = "^8.1.3"
black = {extras = ["jupyter"], version = "^24.4.2"}
fire = "^0.6.0"
jupytext = "^1.16.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
